better type hints

+++ pp/components/extension.py
@@ -1,6 +1,6 @@
 from typing import List, Optional, Callable
 import numpy as np
-from numpy import float64, ndarray
+from numpy import ndarray
 from pp.container import container
 from pp.component import Component
 import pp
@@ -28,7 +28,7 @@ def line(
     return [p0, p1, p2, p3]
 
 
-def move_polar_rad_copy(pos: ndarray, angle: float64, length: float) -> ndarray:
+def move_polar_rad_copy(pos: ndarray, angle: float, length: float) -> ndarray:
     c = np.cos(angle)
     s = np.sin(angle)
     return pos + length * np.array([c, s])
